% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.cv.ensembleEN}
\alias{coef.cv.ensembleEN}
\title{Extract coefficients from a cv.ensembleEN object.}
\usage{
\method{coef}{cv.ensembleEN}(object, index = object$index_opt, ...)
}
\arguments{
\item{object}{Fitted cv.ensembleEN object.}

\item{index}{Indices indicating values of lambda_S at which to extract coefficients. Defaults to the optimal value.}

\item{...}{Additional arguments for compatibility}
}
\value{
A vector of coefficients
}
\description{
Extract coefficients from a cv.ensembleEN object.
}
\seealso{
\code{\link{cv.ensembleEN}}
}
