% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.ensembleEN}
\alias{coef.ensembleEN}
\title{Extract coefficients from an ensembleEN object.}
\usage{
\method{coef}{ensembleEN}(object, index = object$index_opt, ...)
}
\arguments{
\item{object}{Fitted ensembleEN object.}

\item{index}{Indices indicating values of lambda_S at which to predict. Defaults to the optimal value.}

\item{...}{Additional arguments for compatibility}
}
\value{
A vector of coefficients
}
\description{
Extract coefficients from an ensembleEN object.
}
\seealso{
\code{\link{ensembleEN}}
}
